Rename module to match all-lower-case naming convention.
